+++ app/Category.php
@@ -12,4 +12,9 @@ class Category extends Model
     {
         return $this->hasMany(User::class);
     }
+
+    public function jobs()
+    {
+        return $this->hasMany(Job::class, 'category_id');
+    }
 }