+++ app/Http/Controllers/ProfileController.php
@@ -44,24 +44,9 @@ class ProfileController extends Controller
         return view('userprofile.editprofile.personal', compact('cities'));
     }
 
-    public function task()
-    {
-        return view('services.task');
-    }
-
     public function auth()
     {
         return view('userprofile.editprofile.auth');
-    }
-
-    public function serviceList()
-    {
-        return view('services.service-list');
-    }
-
-    public function serviceSingle()
-    {
-        return view('services.service-single');
     }
 
     /**